Extract ambiguous configuration selection error into its own exception type
